requirements.txt
pathlib==1.0.1
sqlparse==0.2.2
psycopg2==2.6.2
uWSGI==2.0.14
PyYAML==3.12
dataset==0.8.0
qvarn==0.81
WebOb==1.7.2
Jinja2==2.9.6
